Separate endpoints into different routes
/category is for get and post, /category/id for put and delete
